Runway_backend/requirements.txt
amqp
asgiref
async-timeout
attrs
autobahn
Automat
beautifulsoup4
billiard
celery
certifi
cffi
channels
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cligj
cloudinary
colorama
constantly
cron-descriptor
cryptography
daphne
Django
django-celery-beat
django-celery-results
django-cloudinary-storage
django-cors-headers
django-environ
django-filter
django-geolocation-fields
django-location-field
django-timezone-field
djangorestframework
djangorestframework-simplejwt
drf-spectacular==0.26.5
drf-yasg==1.21.7
docopt
Fiona
# GDAL
# GDAL @ file:///C:/Users/nihal/pipwin/GDAL-3.4.3-cp310-cp310-win_amd64.whl#sha256=9893da5532a6c5dadc1fc3bc402ae29d161b7cf190d778f129da84d0e50baa29
geographiclib
geopy
gunicorn
hyperlink
idna
incremental
Js2Py
kombu
Markdown
networkx
packaging
Pillow
prompt-toolkit
psycopg2
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
pygeos
pyjsparser
PyJWT
pyOpenSSL
pyotp
PyPrind
pyproj
PyYAML==6.0.1
pySmartDL
python-crontab
python-dateutil
pytz
redis
requests
service-identity
shapely
six
soupsieve
sqlparse
# Twisted==22.10.0
# twisted-iocpsupport==1.0.4
txaio
typing_extensions
tzdata
tzlocal
urllib3
uritemplate==4.1.1
vine
wcwidth
zope.interface
